Feat: questionComment 생성 수정 삭제 구현.

+++ client/src/components/js/comment/Comment.js
@@ -2,8 +2,11 @@ import axios from 'axios';
 import React from 'react'
 import '../../css/comment/Comment.scss'
 import createdAt from '../createdAt/CreatedAt';
+import { useParams, useNavigate } from 'react-router-dom';
 
 export default function Comment({id, content, setEditClick, type, setEditValue, setCommentId, accessToken}) {
+  let params  = useParams();
+
   axios.defaults.headers.common["Authorization"] = accessToken;
 
 
@@ -22,23 +25,29 @@ export default function Comment({id, content, setEditClick, type, setEditValue, 
   // click delete
   const handleDelete = () => {
     if(type === 'question'){
-      // fetchDelete("/question/{id}/comments/{comment-id}")
-      axios.delete(`/api/question/${id}/comments/${content.questionCommentId}`)
-      .then((res) => {
-        console.log(res)
-      })
-      .catch(error => {
-        console.log(error.response);
-      });
+      if (window.confirm("Are you sure you want to delete the comment?") === true) {
+        axios.delete(`/api/question/${id}/comments/${content.questionCommentId}`)
+        .then((res) => {
+          if(res.status === 200) {
+            window.location.href = `/questions/${params.id}`;
+          }
+        })
+        .catch(error => {
+          alert(error.response.data.errors[0].reason);
+        });
+      }
     }else if(type === 'answer'){
-      // fetchDelete("question/{id}/answer/{answer-id}/comments/{comment-id}")
-      axios.delete(`/api/question/${id}/answer/${content.answerCommentId}comments/${content.answerCommentId}`)
-      .then((res) => {
-        console.log(res)
-      })
-      .catch(error => {
-        console.log(error.response);
-      });
+      if (window.confirm("Are you sure you want to delete the comment?") === true) {
+        axios.delete(`/api/question/${id}/answer/${content.answerCommentId}comments/${content.answerCommentId}`)
+        .then((res) => {
+          if(res.status === 200) {
+            window.location.href = `/questions/${params.id}`;
+          }
+        })
+        .catch(error => {
+          alert(error.response.data.errors[0].reason);
+        });
+      }
     }
   }
   return (
